Beta release ver-0.1.1

+++ samarth-hegde/ecommerce.py
@@ -2,24 +2,19 @@ import os
 
 
 def user_delete_request(user_name, user_password):
-    os.remove(f'{user_name}.txt')
-    find_user = f"'{user_name}': '{user_password}'"
-    find_user = '{' + find_user
-    with open('user_registration_data.txt', 'a+') as user_credential:
-        user_credential.seek(0)
-        for data in user_credential:
-            for name in data.split('}'):
-                if find_user == name:
-                    name.replace(name, '')
-                    return True
-    return False
+    if os.path.exists(f'{user_name}.txt'):
+        os.remove(f'{user_name}.txt')
+        os.remove(f'{user_name}_cart.txt')
+        return True
+    else:
+        return False
 
 
 def user_cart(user_name, user_password):
     try:
         user_cart_item = {}
         user_activity_status = True
-        user_cart_data = open(f'{user_name}.txt', 'a+')
+        user_cart_data = open(f'{user_name}_cart.txt', 'a+')
         while user_activity_status:
             print("Please select only respective numbers.")
             print("1.Add to cart 2.Delete from cart 3.Modify quantity 4.View Cart 5.Logout 6.Delete my Account \n")
@@ -28,44 +23,45 @@ def user_cart(user_name, user_password):
                 item_name = input("Enter item name: \t")
                 item_quantity = input(f"Enter {item_name} quantity:\t ")
                 user_cart_item[item_name] = item_quantity
-                print("Your cart has:")
                 for key, value in user_cart_item.items():
-                    print(f"{key} : {value}")
-                    print()
+                    user_cart_data.write(key+' : '+value+'\n')
+                print("Your cart has:")
+                user_cart_data.seek(0)
+                print(user_cart_data.read())
 
             elif user_activity == 2:
                 item_name = input("Enter item name to delete:\t")
 
                 if item_name in user_cart_item.keys():
                     del user_cart_item[item_name]
+                    print(f'{item_name} deleted!')
+                    for key, value in user_cart_item.items():
+                        user_cart_data.write(key + ' : ' + value + '\n')
                 else:
                     print("Item not found!")
-                print(f'{item_name} deleted!')
 
                 print("Your cart has:")
-                for key, value in user_cart_item.items():
-                    print(f"{key} : {value}")
-                    print()
+                user_cart_data.seek(0)
+                print(user_cart_data.read())
 
             elif user_activity == 3:
                 item_name = input("Enter item name to modify.\t")
                 item_quantity = input("Enter quantity to update.\t")
                 updated_item_set = {item_name: item_quantity}
                 user_cart_item.update(updated_item_set)
-                print("Your cart has:")
                 for key, value in user_cart_item.items():
-                    print(f"{key} : {value}")
-                    print()
+                    user_cart_data.write(key + ' : ' + value + '\n')
+                print("Your cart has:")
+                user_cart_data.seek(0)
+                print(user_cart_data.read())
 
             elif user_activity == 4:
-                print("Your cart has:\n")
-                for key, value in user_cart_item.items():
-                    print(f"{key} : {value}")
-                    print()
+                print("Your cart has:")
+                user_cart_data.seek(0)
+                print(user_cart_data.read())
 
             elif user_activity == 5:
                 user_activity_status = False
-                user_cart_data.write(str(user_cart_item))
                 user_cart_data.close()
 
             elif user_activity == 6:
@@ -83,33 +79,38 @@ def user_cart(user_name, user_password):
 
 
 def user_registration():
-    user_registration_data = open('user_registration_data.txt', 'a')
     user_name = str(input("Please Enter user user name: \t"))
     user_password = input("Please Enter password:\t")
-    user_credential = {user_name: user_password}
-    user_registration_data.write(str(user_credential))
-    user_registration_data.close()
-    return True
+    if os.path.exists(f'{user_name}.txt'):
+        print(f'{user_name} is already registered, Please login!!')
+        return False
+    else:
+        user_registration_data = open(f'{user_name}.txt', 'w')
+        # user_credential = {user_name: user_password}
+        user_registration_data.write(user_password)
+        user_registration_data.close()
+        return True
 
 
 def user_login():
     user_name = input("Please enter user name:\t")
     user_password = input("Please enter password:\t")
-    find_user = f"'{user_name}': '{user_password}'"
-    find_user = '{'+find_user
+    # find_user = f"'{user_name}': '{user_password}'"
+    # find_user = '{'+find_user
     user_name_match = False
     user_password_match = False
-    with open('user_registration_data.txt', 'a+') as user_credential:
-        user_credential.seek(0)
-        for data in user_credential:
-            for name in data.split('}'):
-                if find_user == name:
+    if os.path.exists(f'{user_name}.txt'):
+        with open(f'{user_name}.txt','r') as user_credintial:
+            user_credintial.seek(0)
+            for data in user_credintial:
+                if user_password == data:
                     user_name_match = True
                     user_password_match = True
                     break
                 else:
                     user_name_match = False
                     user_password_match = False
+
     if user_password_match and user_name_match:
         print(f"Welcome! {user_name}!")
         user_cart(user_name,user_password)
